rename Inputs.jsx to more explicit Selection.jsx
