+++ frontend/quickcheck/src/components/Selection.jsx
